pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "LangChain demonstration project"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-classic>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
]
